Prevent documents with embedded associations from persisting multiple times for a single save.
Previously, the run_callbacks propagation passed the block on to the run_callbacks call on each embedded document.  The persistence to riak happens as part of this, so that would cause the document to be persisted many times.

Instead, the block should only run _once_, as part of the run_callbacks invocation on the parent document.  The `super` call takes care of passing the block along properly.