Add #replace_links method to linked proxy.
This allows the links to be replaced directly rather than needing to replace them by replacing their corresponding documents.  This is necessary for the linked association conflict resolution (coming in the next commit) so that it can assign the value of the linked association on the resolving document without needing to link walk and load the documents themselves.